+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii 2 Application Runtime Configuration extension Change Log
 
 - Bug: Fixed `Item::attributes()` returns incorrect attributes list (klimov-paul)
 - Enh #7: `Item::$label` converted into virtual property, allowing its automatic generation from `Item::$id` value (nexen2, klimov-paul)
+- Enh #8: Added ability to clear only single item value via `Manager::clearValue()` (klimov-paul)
 
 
 1.0.2, June 3, 2016